feat: refactor provider
