+++ packages/Cargo.toml
@@ -4,6 +4,7 @@ resolver = "2"
 members = [
     "whisky-common",
     "whisky-js",
+    "whisky-provider",
     "whisky-js/json-gen",
     "whisky-csl",
     "whisky",
