+++ packages/whisky-provider/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "whisky"
+name = "whisky-provider"
 version = "1.0.1"
 edition = "2021"
 license = "Apache-2.0"
@@ -9,26 +9,15 @@ description = "The Cardano Rust SDK, inspired by MeshJS"
 crate-type = ["cdylib", "rlib"]
 
 [dependencies]
-getrandom = { version = "0.2", features = ["js"] }
 hex = "0.4"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
-cryptoxide = "0.4.4"
-serde-wasm-bindgen = "0.6.5"
 async-trait = "0.1.79"
 uplc = "=1.1.9"
-wasm-bindgen = { version = "=0.2.92", features = ["serde-serialize"] }
-rand_os = "0.1"
-noop_proc_macro = "0.3.0"
-pallas-codec = { version = "0.30.2", features = ["num-bigint"] }
-pallas-primitives = "0.31.0"
-pallas-traverse = "0.31.0"
 maestro-rust-sdk = "1.1.3"
 whisky-csl = { version = "1.0.1", path = "../whisky-csl" }
 whisky-common = { version = "1.0.1", path = "../whisky-common" }
-whisky-provider = { version = "1.0.1", path = "../whisky-provider" }
 reqwest = "0.12.5"
-tokio = { version = "1.38.0", features = ["macros", "rt-multi-thread"] }
 futures = "0.3.31"
 
 [profile.release]
@@ -40,3 +29,4 @@ lto = true
 
 [dev-dependencies]
 dotenv = "0.15.0"
+tokio = { version = "1.38.0", features = ["macros", "rt-multi-thread"] }
