+++ packages/whisky-provider/src/maestro/models/asset.rs
