+++ packages/whisky-provider/src/maestro/utils/transaction_utils.rs
@@ -1,4 +1,4 @@
-use crate::providers::maestro::models::transaction::TransactionDetail;
+use crate::maestro::models::transaction::TransactionDetail;
 
 use whisky_common::{models::TransactionInfo, WError};
