+++ packages/whisky-provider/tests/blockfrost/fetcher.rs
@@ -2,7 +2,8 @@
 mod tests {
     use dotenv::dotenv;
     use std::env::var;
-    use whisky::{BlockfrostProvider, Fetcher};
+    use whisky_common::Fetcher;
+    use whisky_provider::BlockfrostProvider;
 
     #[tokio::test]
     async fn test_fetch_account_info() {
