+++ packages/whisky/tests/mod.rs
@@ -1,2 +1,1 @@
-mod providers;
 mod transaction;
